fix(eks): rename the lambda-stack cdk script file
