+++ eks/cdk/app.py
@@ -3,7 +3,7 @@ from ssm import DiskStressDocument
 from cw import CloudWatchAlarm
 from eks import EKS
 from fis import FIS
-from eb import LambdaCron
+from fn import LambdaCron
 
 props = {'namespace': 'fisworkshop'}
 app = App()
@@ -12,6 +12,6 @@ ssm = DiskStressDocument(app, f"{props['namespace']}-ssm", props)
 eks = EKS(app, f"{props['namespace']}-eks", props)
 alarm = CloudWatchAlarm(app, f"{props['namespace']}-cw", eks.outputs)
 fis = FIS(app, f"{props['namespace']}-fis", alarm.outputs)
-cron = LambdaCron(app, f"{props['namespace']}-eb", fis.outputs)
+cron = LambdaCron(app, f"{props['namespace']}-lambda", fis.outputs)
 
 app.synth()
